Overhaul of player polling
Changed from a static 1-second poll to using long polling for getting updates from the server, with calls to a new /api/update endpoint to update the server. This also allows the server to know when the player is unloaded.

Also added a slide up animation to the song info overlay.